### 2022-08-09 #4645 @jingpengw
could you also provide the github repo link?

Best,
Jingpeng

On Tue, Aug 9, 2022 at 2:01 AM Øystein Sørensen ***@***.***>
wrote:

> 👋 @jingpengw <https://github.com/jingpengw> @abhi-glitchhg
> <https://github.com/abhi-glitchhg> @surajpaib
> <https://github.com/surajpaib> would any of you be willing to review this
> submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
