### 2022-08-10 #4645 @jingpengw
I am happy to review it, but there exists a package, called
pytorch_connectomics, that is pretty well made. By quick browsing of this
repo, I did not find the uniqueness of this repo compared to it.
https://github.com/zudi-lin/pytorch_connectomics

Best,
Jingpeng

On Tue, Aug 9, 2022 at 8:36 AM Øystein Sørensen ***@***.***>
wrote:

> @jingpengw <https://github.com/jingpengw>, the GitHub repo is
> https://github.com/DiamondLightSource/volume-segmantics.
>
