### 2022-08-10 #4645 @osorensen
> I am happy to review it, but there exists a package, called pytorch_connectomics, that is pretty well made. By quick browsing of this repo, I did not find the uniqueness of this repo compared to it. https://github.com/zudi-lin/pytorch_connectomics Best, Jingpeng
> […](#)
> On Tue, Aug 9, 2022 at 8:36 AM Øystein Sørensen ***@***.***> wrote: @jingpengw <https://github.com/jingpengw>, the GitHub repo is https://github.com/DiamondLightSource/volume-segmantics. — Reply to this email directly, view it on GitHub <[#4645 (comment)](https://github.com/openjournals/joss-reviews/issues/4645#issuecomment-1209325833)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AB2MB5N4AWZQNUWY44Z4GN3VYJGEHANCNFSM55QBZ5AA> . You are receiving this because you were mentioned.Message ID: ***@***.***>

@OllyK could you please provide a comment to this issue?
