### 2022-08-10 #4645 @OllyK
@jingpengw @osorensen Yes certainly. I was just holding off since the invitiation to review hasn't been accepted yet and was unsure if reacting to pre-judgement before the process has started would be constructive 😃
Thanks for pointing out the `pytorch_connectectomics` repo Jingpeng, it certainly looks impressive and appears to have a strong team behind it. Having just looked at it briefly, there are a number of differences that I can see.

1. `pytorch_connectomics` is much more complicated, both in the functionality it offers and the useage. There are many options to configure and the commandline invocations appear to be complex. My package focusses on simplicity and has one main purpose - fast and straightforward training of models for volumetric semantic segmentation with minimal configuration.
2. In `pytorch_connectomics` 2D or 3D models are generally trained from scratch, a time a resource heavy process requiring large training datasets, hence why their examples often distribute training across multiple large GPUs. My library is intended to be used in situations where the user doesn't have access to these resources and does not want to wait much longer than 5 minutes to train a model - the library has been incorporated into a GUI application (SuRVoS2) where users (people completely unfamiliar with computer science) expect to be able to perform steps in a reasonable timeframe. Also we want syncrotron users to quickly train models for use in their current experiments.
3. Although the authors of `pytorch_connectomics` state that their package can be used on all kinds of data their focus appears to be heavily weighted towards EM data and segmentation techniques useful for connectomics. My package is generic to volumetric image data from any modality and discipline.
4. Swapping out and trying different pre-trained encoder types and models appears to be more straightforward in my library, simply change a string in a YAML file. This is thanks to functionality afforded by the `segmentation-models.pytorch` library that is utilised in my package. Therefore a smaller, lightweight encoder could be used to save resources, for example. Perhaps I haven't looked in enough detail but couldn't see the option to change encoder in `pytorch_connectomics`.
5. As previously mentioned, it is hoped that the simple API for `volume_segmantics` would allow the package to be incorparted into other applications (as it is in SuRVoS2). This does not appear to be such a consideration for `pytorch_connectomics`.

Apologies for the long answer, I expect I'll wake up in the middle of the night with some other reasons on my mind 😄  These are the only comments I can make, to the best of my knowledge since I haven't had time to try the `pytorch_connectomics`  package out yet and am unlikely to do so in the near future,

Best Wishes,

Olly
