### 2022-08-21 #4691 @editorialbot
[REVIEW]: Volume Segmantics: A Python Package for Semantic Segmentation of Volumetric Data Using Pre-trained PyTorch Deep Learning Models
**Submitting author:** @OllyK (<a href="http://orcid.org/0000-0002-6152-7207">Oliver N. F. King</a>)
**Repository:** https://github.com/DiamondLightSource/volume-segmantics
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.7
