### 2022-09-09 #4691 @jingpengw
@osorensen
@OllyK
It seems that the `State of the field` is missing. It would be better to compare with existing tools including my own:
https://github.com/flatironinstitute/neutorch

I have also listed some others in the README of my repo.
[DeepEM](https://github.com/seung-lab/DeepEM)
[DataProvider3](https://github.com/seung-lab/DataProvider3)
[PyTorchUtils](https://github.com/nicholasturner1/PyTorchUtils)
[pytorch_connectomics](https://github.com/zudi-lin/pytorch_connectomics)
