### 2022-09-14 82c08de
Paper edits to include state of the field
