+++ paper.bib
@@ -13,7 +13,6 @@
 	month = dec,
 	year = {2015},
 	pages = {211--252},
-	file = {Full Text:C\:\\Users\\vvw07985\\Zotero\\storage\\PW86KSMI\\Russakovsky et al. - 2015 - ImageNet Large Scale Visual Recognition Challenge.pdf:application/pdf},
 }
 
 @inproceedings{perslev_one_2019,
@@ -31,7 +30,6 @@
 	editor = {Shen, Dinggang and Liu, Tianming and Peters, Terry M. and Staib, Lawrence H. and Essert, Caroline and Zhou, Sean and Yap, Pew-Thian and Khan, Ali},
 	year = {2019},
 	pages = {30--38},
-	file = {Submitted Version:C\:\\Users\\vvw07985\\Zotero\\storage\\ZHURWZB7\\Perslev et al. - 2019 - One Network to Segment Them All A General, Lightw.pdf:application/pdf},
 }
 
 @article{buslaev_albumentations_2020,
@@ -74,7 +72,6 @@
 	note = {Publisher: Royal Society},
 	keywords = {computed tomography, contrast agent, flow network, human placenta, machine-learning segmentation, spatial statistics},
 	pages = {20210140},
-	file = {Full Text PDF:C\:\\Users\\vvw07985\\Zotero\\storage\\LCAP6DDP\\Tun et al. - A massively multi-scale approach to characterizing.pdf:application/pdf},
 	month = jun,
 	year = {2021},
 }
@@ -107,4 +104,112 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/DiamondLightSource/SuRVoS2}
+}
+
+@misc{perslev_github_2019,
+  author = {Perslev, Mathias and Igel, Christian},
+  title = {Multi-Planar U-Net},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/perslev/MultiPlanarUNet}
+}
+
+@misc{tekawade_github_2020,
+  author = {Tekawade, Aniket and Igel, Christian},
+  title = {CTSegNet},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/aniketkt/CTSegNet}
+}
+
+@misc{wolny_github_2019,
+  author = {Wolny, Adrian},
+  title = {pytorch-3dunet},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/wolny/pytorch-3dunet}
+}
+
+@misc{chen_rethinking_2017,
+	title = {Rethinking {Atrous} {Convolution} for {Semantic} {Image} {Segmentation}},
+	url = {http://arxiv.org/abs/1706.05587},
+	doi = {10.48550/arXiv.1706.05587},
[…]
