### 2022-09-14 88b2801
Update paper.md
Minor changes around citation placement (not sure but i like this more as it highlights that they are package references) and some slight rewording around the connectomics section
