### 2022-09-14 2e232f1
Minor typo fixes to paper
