+++ paper.md
@@ -86,7 +86,7 @@ will be included such as fusion models [@perslev_one_2019]. A schematic of the t
 
 Currently there are a number of other software implementations available for segmentation of 3D data. Some of these also use 2d networks and combine prediction outputs to 3D, for example the `Multi-Planar U-Net` package [@perslev_github_2019] and the `CTSegNet` package [@tekawade_github_2020]. However, neither of these packages allows the use of pre-trained encoders or multiple model architectures. In addition, the general purpose `pytorch-3dunet` package [@wolny_github_2019] exists to allow training a 3D U-Net on image data, again without the time and resource advantages of pre-trained 2D networks. 
 
-In the field of connectomics, several packages [@lee_deepem_2018; @lin2021pytorch; @urakubo_uni-em_2019; @wu_neutorch_2021] enable the segmentation of structures within the brain, often from electron microscopy data, these could in principle be used in a subject and method-agnostic manner similarly to `volume-segmantics`. One member of this set, the package `pytorch-connectomics`, [@lin_pth_connec_github_2019] allows training of 2D and 3D networks for segmentation of 3D image data as well as customisable strategies for data augmentation and multi-task and semi-supervised learning. Despite the versatility of this software, its deliberate focus on connectomics which is essential for effectiveness in this complex field, mean that there are added levels of complexity for the generalist user. This specialisation also means that there is only one pre-trained 2D model architecture available, and some of the configuration and command-line options are context specific.
+In the field of connectomics, several packages [@lee_deepem_2018; @lin2021pytorch; @urakubo_uni-em_2019; @wu_neutorch_2021] enable the segmentation of structures within the brain, often from electron microscopy data, these could in principle be used in a subject and method-agnostic manner similarly to `volume-segmantics`. One member of this set, the package `pytorch-connectomics`, [@lin_pth_connec_github_2019] allows training of 2D and 3D networks for segmentation of 3D image data as well as customisable strategies for data augmentation and multi-task and semi-supervised learning. Despite the versatility of this software, its deliberate focus on connectomics which is essential for effectiveness in this complex field, means that there are added levels of complexity for the generalist user. This specialisation also means that there is only one pre-trained 2D model architecture available, and some of the configuration and command-line options are context specific.
 
 ## Real-world usage
 
