### 2022-09-22 #4691 @OllyK
Hi @osorensen , according to the docs, only editors can run the editorialbot `check references` command. Would it be possible for you to run it? -  I updated the references last week. Kind regards
