### 2022-10-03 #4691 @OllyK
@estenhl  Feel free to ask if there's anything that needs clarifying or updating. The paper doesn't contain any original results, it references studies where we've used earlier versions of the code to segment data. Best Wishes.
