### 2022-10-04 #4691 @estenhl
Sorry for the late reply @osorensen, the missing check was purely due to a mistake on my part. @OllyK, great work, and thanks for the detailed replies to @jinpengw's questions, both here and in the pre-review thread
