### 2022-10-04 #4691 @OllyK
Thanks very much @osorensen ! And once again many thanks to @jingpengw and @estenhl for taking the time to review

> @OllyK, I will now read through the manuscript a final time, and let you know if I have any suggested changes. Meanwhile, could you

 - [x ]  Make a tagged release of your software, and list the version tag of the archived version here.
 - [x ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
 - [x ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
 - [x ]  Please list the DOI of the archived version here.
The DOI of the archive is [https://doi.org/10.5281/zenodo.7143363](https://doi.org/10.5281/zenodo.7143363)

Please let me know if there's anything else you need.
Best Wishes,

OllyK
