### 2022-10-04 DiamondLightSource/volume-segmantics#16 @osorensen
Editorial comments on JOSS submission: Reference to PyTorch
I'll post some comments on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4691) as issues in this repo. This is the first one.

On line 24, please include a reference to PyTorch. It can be found here: https://github.com/pytorch/pytorch/blob/master/CITATION
