### 2022-10-04 DiamondLightSource/volume-segmantics#17 @osorensen
Editorial comment to JOSS submission
Here's another comment to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4691):

On line 50, please change "2d" to "2D", i.e., use uppercase D to be consistent.
