### 2022-10-04 DiamondLightSource/volume-segmantics#18 @osorensen
Editorial comment to JOSS submission
Another comment to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4691).

On line 59 you write `volume-segmantics`. Do you here refer to `Volume Segmantics`, i.e., the package being submitted?
