### 2022-10-04 DiamondLightSource/volume-segmantics#19 @osorensen
Editorial comment to JOSS submission
Another comment, https://github.com/openjournals/joss-reviews/issues/4691.

On line 73, please make sure the word "pixels" appears in normal font (not *italic*), and add a space between "2120" and "pixels".
