### 2022-10-04 DiamondLightSource/volume-segmantics#20 @osorensen
Editorial comment to JOSS submission
Another comment to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4691).

Could you please elaborate what you mean by the following sentence? In particular, where is "here"?

> These models can then be used to segment data during their time here and also when back at their home institution.

https://github.com/DiamondLightSource/volume-segmantics/blob/a638ebf7c6066f09b419a76432f3e2b0fe09f43a/paper.md?plain=1#L115-L117
