### 2022-10-04 0c51804
Adds reference to PyTorch
- Resolves #16
