### 2022-10-04 DiamondLightSource/volume-segmantics#21 @osorensen
References
Some comment to the references in the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4691).

- Please add a DOI for this reference: https://arxiv.org/abs/2112.05754
- Please make sure Multi-Planar U-Net has correct capitalization. This probably requires adding curly braces around the title in the BibTeX source.
