### 2022-10-04 #4691 @osorensen
@OllyK, I think the paper is clear and well written. As you see, I have opened some rather small issues in the source repository. Please let me know when they are fixed (or if you disagree with my suggestions, which is absolutely fine), and I can move forward with accepting the submission.
