### 2022-10-04 facd828
Fixes errors in references
- Fixes capitalisation of Multi-Planar U-Net
- Adds DOI for pytorch-connectomics paper
- Resolves #21
