+++ paper.bib
@@ -108,7 +108,7 @@
 
 @misc{perslev_github_2019,
   author = {Perslev, Mathias and Igel, Christian},
-  title = {Multi-Planar U-Net},
+  title = {Multi-Planar {U}-Net},
   year = {2019},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -172,7 +172,8 @@ version: 3},
   title={PyTorch Connectomics: A Scalable and Flexible Segmentation Framework for EM Connectomics},
   author={Lin, Zudi and Wei, Donglai and Lichtman, Jeff and Pfister, Hanspeter},
   journal={arXiv preprint arXiv:2112.05754},
-  year={2021}
+  year={2021},
+  doi={10.48550/arXiv.2112.05754}
 }
 
 @article{urakubo_uni-em_2019,
