### 2022-10-04 12a4457
Clarifies text to elaborate on location
- Resolves #20
