### 2022-10-04 b7b152a
Clarifies text to elaborate on location
- Resolves #20
