### 2022-10-05 02a665b
Fixes capitalisation of 2D
- Resolves #17
