### 2022-10-05 16ebeff
Fixes capitalisation of 2D
- Resolves #17
