### 2022-10-05 c284d23
Fixes package name
- Resolves #18
