### 2022-10-05 6cd3ac9
Fixes italic font error
- Resolves #19
