### 2022-10-05 153b566
Fixes italic font error
- Resolves #19
