### 2022-10-05 77543bc
Bumps version to 0.2.7 in pyproject.toml and README
