### 2022-10-09 #4691 @arfon
@OllyK – the link works for me? https://joss.theoj.org/papers/by/Oliver%20N%2E%20f%2E%20King

What browser are you using out of interest?
