### 2022-10-09 #4691 @OllyK
> @OllyK – the link works for me? https://joss.theoj.org/papers/by/Oliver%20N%2E%20f%2E%20King
>
> What browser are you using out of interest?
@arfon Strangely the link works in Firefox but doesn't work in Chrome (on Windows) or on a DuckDuckGo browser on my Android phone. Like I said, not a big deal, happy it's been published 😄. Thanks @osorensen @estenhl and @jingpengw for taking the time to review!
![image](https://user-images.githubusercontent.com/11923466/194757304-5e5883ee-0335-45b4-9e85-e4794c85c1ad.png)
